backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
openai>=1.0.0
anthropic>=0.20.0
google-generativeai